simulator/requirements.txt
kafka-python
river
